+++ src/main/kotlin/org/wfanet/measurement/common/db/postgres/PostgresFlags.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.postgres
+package org.wfanet.measurement.common.db.postgres
 
 import picocli.CommandLine
