+++ src/main/kotlin/org/wfanet/measurement/gcloud/postgres/BUILD.bazel
@@ -9,3 +9,13 @@ kt_jvm_library(
     srcs = ["PostgresFlags.kt"],
     deps = ["//imports/java/picocli"],
 )
+
+kt_jvm_library(
+    name = "options",
+    srcs = ["PostgresConnectionFactoryOptions.kt"],
+    deps = [
+        "//imports/java/com/google/cloud/sql:r2dbc-core",
+        "//imports/java/org/postgresql:r2dbc",
+        "//src/main/kotlin/org/wfanet/measurement/gcloud/postgres:flags",
+    ],
+)
