+++ src/main/kotlin/org/wfanet/measurement/gcloud/spanner/tools/BUILD.bazel
@@ -9,6 +9,7 @@ kt_jvm_library(
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/common",
         "//src/main/kotlin/org/wfanet/measurement/common/db/liquibase",
+        "//src/main/kotlin/org/wfanet/measurement/common/db/liquibase/tools:update_schema",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/spanner:flags",
     ],
 )
